Remove AddPeerProtocol
