+++ rpcserver.go
@@ -12,6 +12,9 @@ import (
 	"time"
 
 	pbrpc "github.com/ethresearch/sharding-p2p-poc/pb/rpc"
+	peer "github.com/libp2p/go-libp2p-peer"
+	pstore "github.com/libp2p/go-libp2p-peerstore"
+	ma "github.com/multiformats/go-multiaddr"
 	"google.golang.org/grpc"
 
 	logging "github.com/ipfs/go-log"
@@ -24,6 +27,33 @@ type server struct {
 	node              *Node
 	serializedSpanCtx []byte
 	rpcServer         *grpc.Server
+}
+
+func parseAddr(addrString string) (peer.ID, ma.Multiaddr, error) {
+	// The following code extracts target's the peer ID from the
+	// given multiaddress
+	ipfsaddr, err := ma.NewMultiaddr(addrString) // ipfsaddr=/ip4/127.0.0.1/tcp/10000/ipfs/QmVmDaabYcS3pn23KaFjkdw6hkReUUma8sBKqSDHrPYPd2
+	if err != nil {
+		return "", nil, err
+	}
+
+	pid, err := ipfsaddr.ValueForProtocol(ma.P_IPFS) // pid=QmVmDaabYcS3pn23KaFjkdw6hkReUUma8sBKqSDHrPYPd2
+	if err != nil {
+		return "", nil, err
+	}
+
+	peerid, err := peer.IDB58Decode(pid) // peerid=<peer.ID VmDaab>
+	if err != nil {
+		return "", nil, err
+	}
+
+	// Decapsulate the /ipfs/<peerID> part from the target
+	// /ip4/<a.b.c.d>/ipfs/<peer> becomes /ip4/<a.b.c.d>
+	targetPeerAddr, _ := ma.NewMultiaddr(
+		fmt.Sprintf("/ipfs/%s", peer.IDB58Encode(peerid)),
+	)
+	targetAddr := ipfsaddr.Decapsulate(targetPeerAddr)
+	return peerid, targetAddr, nil
 }
 
 func makeResponse(success bool, message string) *pbrpc.Response {
@@ -66,17 +96,24 @@ func (s *server) AddPeer(
 		return makePlainResponse(false, fmt.Sprintf("Failed to generate peer key/ID with seed: %v", req.Seed)), nil
 	}
 
-	var replyMsg string
-	err = s.node.AddPeer(spanctx, mAddr)
-	success := (err == nil)
-	if success {
-		replyMsg = fmt.Sprintf("Added Peer %v:%v, pid=%v!", req.Ip, req.Port, targetPID)
-		// Tag the span with peer info
-		logger.SetTag(spanctx, "Added peer", fmt.Sprintf("%v:%v", req.Ip, req.Port))
-	} else {
-		replyMsg = fmt.Sprintf("Failed to add Peer %v:%v, pid=%v!", req.Ip, req.Port, targetPID)
+	peerid, targetAddr, err := parseAddr(mAddr)
+	if err != nil {
+		logger.FinishWithErr(spanctx, fmt.Errorf("Failed to parse peer address: %s, err: %v", mAddr, err))
+		log.Printf("Failed to parse peer address: %s, err: %v", mAddr, err)
+		return makePlainResponse(false, fmt.Sprintf("Failed to parse peer address: %s, err: %v", mAddr, err)), nil
 	}
-	return makePlainResponse(success, replyMsg), nil
+	s.node.Peerstore().AddAddr(peerid, targetAddr, pstore.PermanentAddrTTL)
+
+	if err := s.node.Connect(ctx, s.node.Peerstore().PeerInfo(peerid)); err != nil {
+		logger.FinishWithErr(spanctx, fmt.Errorf("Failed to connect to peer %v, err: %v", peerid, err))
+		log.Printf("Failed to connect to peer %v, err: %v", peerid, err)
+		return makePlainResponse(false, fmt.Sprintf("Failed to connect to peer %v, err: %v", peerid, err)), nil
+	}
+
+	// Tag the span with peer info
+	logger.SetTag(spanctx, "Added peer", fmt.Sprintf("%v:%v", req.Ip, req.Port))
+
+	return makePlainResponse(true, fmt.Sprintf("Added Peer %v:%v, pid=%v!", req.Ip, req.Port, targetPID)), nil
 }
 
 func (s *server) SubscribeShard(
